requirements.txt
ffprobe==0.5
ffprobe-python==1.0.3
filetype==1.1.0
jenkspy==0.3.1
joblib==1.2.0
numpy==1.23.4
Pillow==9.2.0
python-dateutil==2.8.2
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
